requirements.txt
pygithub

--editable ./opsutilshelpers
